add tests for Spotlight components
